+++ src/index.ts
@@ -20,7 +20,6 @@ import MentionsFollowedScorer from './scorer/feature/mentions_followed_scorer';
 import MostFavouritedAccountsScorer from "./scorer/feature/most_favourited_accounts_scorer";
 import MostRepliedAccountsScorer from "./scorer/feature/most_replied_accounts_scorer";
 import MostRetootedAccountsScorer from "./scorer/feature/most_retooted_accounts_scorer";
-import NumericFilter from "./filters/numeric_filter";
 import NumFavouritesScorer from "./scorer/feature/num_favourites_scorer";
 import NumRepliesScorer from "./scorer/feature/num_replies_scorer";
 import NumRetootsScorer from "./scorer/feature/num_retoots_scorer";
@@ -63,10 +62,8 @@ import {
     ScorerDict,
     StorageKey,
     StringNumberDict,
-    TrendingLink,
     TrendingObj,
     TrendingStorage,
-    TagWithUsageCounts,
     TrendingWithHistory,
     WeightName,
     Weights,
@@ -595,17 +592,13 @@ export {
     VIDEO_TYPES,
     Account,
     BooleanFilterName,
-    FeedFilterSettings,
     MediaCategory,
     StringNumberDict,
     Toot,
-    TrendingLink,
     TrendingObj,
-    TagWithUsageCounts,
     TrendingWithHistory,
     TypeFilterName,
     WeightName,
-    Weights,
     // Helpers we also export
     extractDomain,
     formatScore,