+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Remove unnecessary exports of TrendingLink, TagWithUsageCounts, FeedFilterSettings, Weights types
 
 ### v0.39.4
 * Remove HTML tags from `Toot.content` field when scanning for hashtags and keywords with `containsString()`